Move action methods to BaseImportModule
